v1/requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
tiktoken>=0.7.0
langchain_text_splitters